# src/main/java/ch/naviqore/api/model/LegType.java
package ch.naviqore.api.model;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

/**
 * Gets or Sets LegType
 */

public enum LegType {
  
  WALK("WALK"),
  
  ROUTE("ROUTE");

  private final String value;

  LegType(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static LegType fromValue(String value) {
    for (LegType b : LegType.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}